Refactor HTML diff generation into separate class
This change extracts HTML diff generation into its own class, HtmlDiffGenerator, from FixCommonErrors form. The main reason for this is to improve code modularity and maintainability. Additionally, it helps to maintain the single responsibility principle by separating the error fixing and HTML diff generation concerns. Now, HTML diff generation is more reusable and testable on its own.
